+++ cpp/opendnp3tests/src/MockPhysicalLayerMonitor.cpp
@@ -77,15 +77,15 @@ void MockPhysicalLayerMonitor::_OnReceive(const uint8_t* apData, size_t aNumByte
 	if(mExpectReadBuffer.Size() < mBytesRead + aNumBytes) {
 		BOOST_FAIL("Read more data than expected");
 	}
-	CopyableBuffer expecting(mExpectReadBuffer + mBytesRead, aNumBytes);
-	CopyableBuffer read(apData, aNumBytes);
+	CopyableBuffer expecting(mExpectReadBuffer + mBytesRead, static_cast<uint32_t>(aNumBytes));
+	CopyableBuffer read(apData, static_cast<uint32_t>(aNumBytes));
 	// check that we're receiving what was written
 	if(expecting != read) {
 		std::ostringstream oss;
 		oss << "Data corruption on receive, " << read << " != " << expecting;
 		BOOST_FAIL(oss.str());
 	}
-	mBytesRead += aNumBytes;
+	mBytesRead += static_cast<uint32_t>(aNumBytes);
 	LOG_BLOCK(LogLevel::Info, "Received " << mBytesRead << " of " << mExpectReadBuffer.Size());
 	WriteBuffer buffer(mReadBuffer, mReadBuffer.Size());
 	mpPhys->AsyncRead(buffer);
@@ -141,8 +141,8 @@ bool MockPhysicalLayerMonitor::AllExpectedDataHasBeenReceived()
 void MockPhysicalLayerMonitor::TransmitNext()
 {
 	if(mWriteBuffer.Size() > this->mBytesWritten) {
-		size_t remaining = mWriteBuffer.Size() - mBytesWritten;
-		size_t toWrite = Min<size_t>(4096, remaining);
+		uint32_t remaining = mWriteBuffer.Size() - mBytesWritten;
+		uint32_t toWrite = Min<uint32_t>(4096, remaining);
 		ReadOnlyBuffer buff(mWriteBuffer + mBytesWritten, toWrite);
 		mpPhys->AsyncWrite(buff);
 		this->mLastWriteSize = toWrite;
